pyproject.toml
[project]
name = "podslurp"
version = "0.1.0"
description = "Search, download and transcribe podcast episodes from the command line"
requires-python = ">=3.9"
dependencies = [
    "faster-whisper>=1.1",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
    "rich>=13",
]

[project.scripts]
podslurp = "podslurp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.4",
]
